[change] Re-organizes component file structure.
