+++ src/components/experiences/InfoTile.tsx
