+++ src/components/experiences/RowDetails.tsx
