+++ src/components/experiences/RowInfo.tsx
